finished Statistics class and now have to finish sendStats() in drone, in order to send stats to the server

+++ src/main/java/Amministrazione/Statistics/Stat.java
@@ -15,7 +15,7 @@ public class Stat {
     public double getDeliveriesCount() {
         return deliveriesCount;
     }
-    public void setDeliveriesCount(int deliveriesCount) {
+    public void setDeliveriesCount(double deliveriesCount) {
         this.deliveriesCount = deliveriesCount;
     }
